Copy Dict.elm from elm-core.
This is the standard Elm implementation of Dict, which
was written mostly by evancz (founder of Elm) and
Skinney.

I will modify this in subsequent commits to expose
some internals, but otherwise I have no other intended
changes.

I pulled this commit:

    https://github.com/elm/core/commit/e001a2a0c90201b760ddbc1dd534a001b359a3e7#diff-5ba3455b19da3db816d65f6add654703

It's from May 2018, but it was in use as recently as
late 2019 for version 19.x of Elm.